feat : add yetTodo, doneTodo counting
